requirements.txt
boto3==1.16.40
botocore==1.19.40
jmespath==0.10.0
pkg-resources==0.0.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.15.0
